Rename evaluators package to evaluator
