+++ neuralogic/nn/evaluator/dynet.py
